[refactor] repository/dto -> repository/query/dto 로 한단계 아래로 이동 - #9
