+++ novelservice/src/main/java/com/iucyh/novelservice/novel/repository/query/dto/NovelSimpleQueryDto.java
@@ -1,4 +1,4 @@
-package com.iucyh.novelservice.novel.repository.dto;
+package com.iucyh.novelservice.novel.repository.query.dto;
 
 import com.iucyh.novelservice.novel.domain.Novel;
 import com.querydsl.core.annotations.QueryProjection;
