+++ novelservice/src/main/java/com/iucyh/novelservice/novel/repository/query/pagingquery/NovelLikeCountPagingQuery.java
@@ -1,7 +1,7 @@
 package com.iucyh.novelservice.novel.repository.query.pagingquery;
 
 import com.iucyh.novelservice.novel.domain.Novel;
-import com.iucyh.novelservice.novel.repository.dto.NovelPagingQueryDto;
+import com.iucyh.novelservice.novel.repository.query.dto.NovelPagingQueryDto;
 import com.iucyh.novelservice.novel.repository.dto.QNovelSimpleQueryDto;
 import com.iucyh.novelservice.novel.repository.query.cursor.NovelCursor;
 import com.iucyh.novelservice.novel.repository.query.cursor.NovelLikeCountCursor;
