+++ novelservice/src/main/java/com/iucyh/novelservice/novel/repository/query/pagingquery/NovelPagingQuery.java
@@ -1,6 +1,6 @@
 package com.iucyh.novelservice.novel.repository.query.pagingquery;
 
-import com.iucyh.novelservice.novel.repository.dto.NovelPagingQueryDto;
+import com.iucyh.novelservice.novel.repository.query.dto.NovelPagingQueryDto;
 import com.iucyh.novelservice.novel.repository.query.cursor.NovelCursor;
 import com.iucyh.novelservice.novel.enumtype.NovelSortType;
 import com.querydsl.jpa.impl.JPAQuery;
