+++ novelservice/src/main/java/com/iucyh/novelservice/novel/repository/query/pagingquery/NovelPagingQueryBaseTemplate.java
@@ -1,6 +1,6 @@
 package com.iucyh.novelservice.novel.repository.query.pagingquery;
 
-import com.iucyh.novelservice.novel.repository.dto.NovelPagingQueryDto;
+import com.iucyh.novelservice.novel.repository.query.dto.NovelPagingQueryDto;
 import com.iucyh.novelservice.novel.repository.query.cursor.NovelCursor;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.dsl.BooleanExpression;
