+++ novelservice/src/main/java/com/iucyh/novelservice/novel/service/NovelQueryService.java
@@ -6,7 +6,7 @@ import com.iucyh.novelservice.novel.dto.mapper.NovelResponseMapper;
 import com.iucyh.novelservice.novel.dto.response.NovelResponse;
 import com.iucyh.novelservice.novel.dto.request.NovelPagingRequest;
 import com.iucyh.novelservice.novel.enumtype.NovelSortType;
-import com.iucyh.novelservice.novel.repository.dto.NovelPagingQueryDto;
+import com.iucyh.novelservice.novel.repository.query.dto.NovelPagingQueryDto;
 import com.iucyh.novelservice.novel.service.codec.NovelCursorBase64Codec;
 import com.iucyh.novelservice.novel.repository.NovelRepository;
 import com.iucyh.novelservice.novel.repository.query.NovelQueryRepository;
